class-json-ingest: Oh goodness, test branch push, useless file.
This abstracts the strings ingest from the JSON into it's own file, but
my lack of Python exposure has me finding that I need to put __init__.py
files in effectively every sub-directory which feels crazy. I'm half
doing this commit to test if I can make a branch, and I'm half doing
this commit to push myself with my ability to squash / make this commit
invisible in the future.